refactor: remove old data model code
