+++ src/core/utils/types.ts
@@ -1,3 +1,9 @@
+/**
+ * App version generated during the build phase.
+ */
+// @ts-ignore: Ignore typing because APP_VERSION is injected by Vite during the build phase.
+export const VERSION = APP_VERSION;
+
 export type UUID = string;
 
 export type UnionOfTupleValues<T extends Readonly<string[] | number[] | symbol[]>> = T[number];
