+++ src/core/validationRules.ts
@@ -1,5 +1,3 @@
-import { Path } from "./model/fileTree";
-
 export const required = (msg: string) => (v: string) => !!v.trim() || msg;
 
 export const counter = (msg: string, count: number) => (v: string) => v.length <= count || msg;
@@ -29,7 +27,8 @@ export const numberInRange =
 export const hex = (msg: string) => (v: string) =>
 	/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/.test(v) || msg;
 
-export const folderName = (msg: string) => (v: string) => !Path.ILLEGAL_CHARS_REGEX.test(v) || msg;
+const ILLEGAL_CHARS_REGEX = /(\/|\\|:|\||<|>|\?|"|\*|%)+/g;
+export const folderName = (msg: string) => (v: string) => !ILLEGAL_CHARS_REGEX.test(v) || msg;
 
 export default {
 	required,
